+++ xpybuild/internal/main.py
@@ -53,6 +53,7 @@ log = logging.getLogger('xpybuild')
 _TASK_BUILD = 'build'
 _TASK_CLEAN = 'clean'
 _TASK_REBUILD = 'rebuild'
+_TASK_LIST_SEARCH = "search"
 _TASK_LIST_TARGETS = 'listTargets'
 _TASK_LIST_FIND_TARGETS = 'findTargets'
 _TASK_LIST_PROPERTIES = 'listProperties'
@@ -65,6 +66,7 @@ def main(args):
 	
 	try:
 		usage = [
+###############################################################################
 '',
 'eXtensible Python-based Build System %s on Python %s.%s.%s'% (XPYBUILD_VERSION, sys.version_info[0], sys.version_info[1], sys.version_info[2]),
 '',
@@ -82,6 +84,8 @@ def main(args):
 '  BUILD_NUMBER=n             Build number string, for reporting and use by build',
 '',
 'Operations: ',
+###############################################################################
+
 '  (if none is specified, the default operation is a normal build)',
 '      --clean                Clean specified targets incl all deps (default=all)',
 '      --rebuild              Clean specified targets incl all deps then build',
@@ -91,9 +95,12 @@ def main(args):
 '                             fast but less correct way to get a quick ',
 '                             incremental build, so use with care. ',
 '',
-' --ft --find-targets <str>   List targets containing the specified substring', 
-' --ti --target-info <str>    Print details including build file location for ',
-'                             targets containing the specified substring',
+'   -s --search <str>         Show info on targets/tags/properties/options ',
+'                             containing the specified substring or regex', 
+# hide these from usage (though they still work), as superceded by the more useful "-s" option
+#' --ft --find-targets <str>   List targets containing the specified substring', 
+#' --ti --target-info <str>    Print details including build file location for ',
+#'                             targets containing the specified substring',
 '      --targets              List available targets and tags (filtered by any ', 
 '                             target or tag names specified on the command line)',
 '      --properties           List properties that can be set and their ',
@@ -162,11 +169,11 @@ def main(args):
 		findTargetsPattern = None
 		format = "default"
 
-		opts,targets = getopt.gnu_getopt(args, "knJh?x:j:l:L:f:F:", 
+		opts,targets = getopt.gnu_getopt(args, "knJh?x:j:l:L:f:F:s:", 
 			["help","exclude=","parallel","workers=","keep-going",
 			"log-level=","logfile=","buildfile=", "dry-run",
 			"targets", 'target-info=', 'ti=', "properties", "options", "clean", "rebuild", "rebuild-ignore-deps", "rid", "ignore-deps", "id",
-			"format=", "timefile=", "ft=", "find-targets=", "depgraph=", 'cpu-stats', 'random-priority', 'profile', 'verify'])
+			"format=", "timefile=", "ft=", "find-targets=", "search=", "depgraph=", 'cpu-stats', 'random-priority', 'profile', 'verify'])
 		
 		for o, a in opts: # option arguments
 			o = o.strip('-')
@@ -185,6 +192,9 @@ def main(args):
 			elif o in ['target-info', 'ti']:
 				task = _TASK_LIST_TARGET_INFO
 				findTargetsPattern = a
+			elif o in ['search', 's']:
+				task = _TASK_LIST_SEARCH
+				searchPattern = a
 			elif o in ['properties']:
 				task = _TASK_LIST_PROPERTIES
 			elif o in ['options']:
@@ -438,6 +448,50 @@ def main(args):
 			for t in findTargetsList:
 				# this must be very easy to copy+paste, so don't put anything else on the line at all
 				print('%s'%(t.name), file=stdout)
+
+		elif task == _TASK_LIST_SEARCH:
+			def showPatternMatches(x): # case sensitive is fine (and probably useful)
+				if searchPattern.replace('\\', '/') in x.replace('\\','/'): return True # basic substring check (with path normalization)
+				if '*' in searchPattern or '?' in searchPattern or '[' in searchPattern: # probably a regex
+					if re.search(searchPattern, x): return True
+				return False
+
+			for t in init.targets().values():
+				t._resolveTargetPath(init)
+
+			print('', file=stdout)
+
+			tagMatches = [t for t in init.tags() if showPatternMatches(t)]
+			if tagMatches: 
+				print ('%d matching tags:'%len(tagMatches), file=stdout)
+				for t in sorted(tagMatches):
+					print(t, file=stdout)
+				print('', file=stdout)
+				
+			targetMatches = [t for t in init.targets().values() if showPatternMatches(t.name) or showPatternMatches(t.path)]
+			if targetMatches: 
+				print ('%d matching targets:'%len(targetMatches), file=stdout)
+				for t in sorted(targetMatches, key=lambda t:(t.type+' '+t.name)):
+					print('- %s priority: %s, tags: [%s]\n   output:  %s\n   defined:  %s'%(t, t.getPriority(), ' '.join(t.getTags()) or 'none', os.path.relpath(t.path), t.location), file=stdout)
+				print('', file=stdout)
+
+			propMatches = {key:value for (key,value) in init.getProperties().items() if showPatternMatches(key)}
+			if propMatches:
+				print('%d matching properties:'%len(propMatches), file=stdout)
+				pad = max(list(map(len, propMatches.keys())))
+				for k in sorted(propMatches.keys()):
+					print(('%'+str(pad)+'s = %s') % (k, propMatches[k]), file=stdout)
+					if init._propertyLocations[k]: # don't do this for built-in property like BUILD_MODE
+						print(('%'+str(pad)+'s   (defined: %s)') % ('', init._propertyLocations[k]), file=stdout)
+					
+			options = init.mergeOptions(None)
+			optionMatches = {key:value for (key,value) in options.items() if showPatternMatches(key)}
+			if optionMatches:
+				print('%d matching options:'%len(optionMatches), file=stdout)
+				pad = max(list(map(len, optionMatches.keys())))
+				for k in sorted(optionMatches.keys()):
+					print(('%'+str(pad)+'s = %s') % (k, optionMatches[k]), file=stdout)
+
 				
 		elif task in [_TASK_BUILD, _TASK_CLEAN, _TASK_REBUILD]: